demo code, fixed adding users

+++ Authentication/routes/api.js
@@ -14,28 +14,42 @@ router.route('/api/users')
 })
 
 .post(function(req, res) {
-    var match = req.body.email.match(/(@bu\.edu)$/);
-    if (match != null) {
-        models.User.findOrCreate({where: {email: req.body.email} }).then(function(success) {
-            res.json({message: 'User created!'});
+    if (!req.user) {
+        res.redirect('/');
+    } else {
+        var match = req.body.email.match(/(@bu\.edu)$/);
+        if (match != null) {
+            models.User.findOrCreate({where: {email: req.body.email} }).then(function(success) {
+                if (!req.user) {
+                    res.redirect('/');
+                } else {
+                    models.User.findAll().then(function(results) {
+                        res.render('admin', {user: req.user, authorized: results});
+                    });
+                }
+            },
+            function(err) {
+                res.send(err);
+            })
+        } else {
+            res.json({message: 'Error: no BU email detected'});
+        }
+    }
+});
+
+router.route('/api/users/:id/del')
+
+.post(function(req,res) {
+    if (!req.user) {
+        res.redirect('/');
+    } else {
+        models.User.destroy({where: {id: req.params.id}}).then(function(success) {
+            res.redirect('/admin');
         },
         function(err) {
             res.send(err);
         })
-    } else {
-        res.json({message: 'Error: no BU email detected'});
     }
-});
-
-router.route('/api/users/:id')
-
-.delete(function(req,res) {
-    models.User.destroy({where: {id: req.params.id}}).then(function(success) {
-        res.json({message: "User deleted!"});
-    },
-    function(err) {
-        res.send(err);
-    })
 });
